weather-assistant/requirements.txt
Flask==2.2.5
requests==2.31.0
spacy==3.6.1
python-dotenv==1.0.0